silence compiler under linux

+++ nsd/tls.c
@@ -2695,7 +2695,7 @@ Ns_TLS_CtxServerCreateCfg(Tcl_Interp *interp,
 
         mem = ns_malloc(sizeof(unsigned int) + (size_t)alpnDs.length);
         *mem = (unsigned int)alpnDs.length;
-        memcpy(mem + 1, alpnDs.string, alpnDs.length);
+        memcpy(mem + 1, alpnDs.string, (size_t)alpnDs.length);
         SSL_CTX_set_alpn_select_cb(ctx, ALPNSelectCB, mem + 1);
         Tcl_DStringFree(&alpnDs);
     }
